requirements.txt
Flask
python-dotenv
langchain-tavily
langchain-groq
langchain-community
langchain-core
streamlit
firebase-admin